requirements.txt
aiohttp
numpy
Pillow
sanic
sanic_cors
